+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Allow disabling the post transform during build (#121)
 * Changed `--skip-install` (default: `false`) flag to `--install` (default: `true`) for scaffolding
 * Changed `--only-core` to `framework` (supporting `piral-base`)
+* Allow async `setup` in pilets
 * Switched to the new Piral logo (#27)
 
 ## 0.9.3 (December 17, 2019)